refactor: 기능 분리 context 로 이동
